Fixed ball getting stuck in mid-air bug
When you pull the paddle away from under the ball, it sometimes gets
stuck in mid-air. I added a check to see if the ball is standing still
for 3 frames after exiting the collision with the paddle, and if so
apply a tiny upward force to get it moving again.